test_tpm2_session: factor out some reusable parts to a shared header
The mock of __wrap_Esys_StartAuthSession can be reused in a unit test for
tpm2_auth_util_get_shandle()
